# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MPF Compare API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MpfFundReturn {
  period: string;
  /** @nullable */
  annualizedReturn: number | null;
  /** @nullable */
  cumulativeReturn: number | null;
}

export interface MpfFundSummary {
  cfId: string;
  nameEn: string;
  /** @nullable */
  nameZh?: string | null;
  trustee: string;
  trusteeCode: string;
  scheme: string;
  fundType: string;
  fundCategory: string;
  /** @nullable */
  launchDate?: string | null;
  /** @nullable */
  fundSizeHkm?: number | null;
  /** @nullable */
  riskClass?: number | null;
  /** @nullable */
  ferPct?: number | null;
  /** @nullable */
  returnAnn?: number | null;
  /** @nullable */
  returnCum?: number | null;
  period: string;
}

export interface MpfFundDetail {
  cfId: string;
  nameEn: string;
  /** @nullable */
  nameZh?: string | null;
  trustee: string;
  trusteeCode: string;
  scheme: string;
  fundType: string;
  fundCategory: string;
  /** @nullable */
  launchDate?: string | null;
  /** @nullable */
  fundSizeHkm?: number | null;
  /** @nullable */
  riskClass?: number | null;
  /** @nullable */
  ferPct?: number | null;
  /** @nullable */
  mgmtFee?: string | null;
  returns: MpfFundReturn[];
}

export interface MpfRankingsResponse {
  period: string;
  asOf: string;
  top: MpfFundSummary[];
  bottom: MpfFundSummary[];
}

export interface MpfCategoryStat {
  category: string;
  /** @nullable */
  avgReturn: number | null;
  fundCount: number;
  period: string;
}

export interface MpfTrusteeStat {
  trustee: string;
  trusteeCode: string;
  /** @nullable */
  avgReturn: number | null;
  fundCount: number;
  period: string;
  funds: MpfFundSummary[];
}

export interface MpfSyncResult {
  message: string;
  count: number;
  /** @nullable */
  error?: string | null;
}

export interface MpfSyncStatus {
  status: string;
  /** @nullable */
  fundsScraped?: number | null;
  /** @nullable */
  errorMessage?: string | null;
  startedAt: string;
  /** @nullable */
  completedAt?: string | null;
}

export type ListMpfFundsParams = {
  period?: string;
  trustee?: string;
  category?: string;
};

export type GetMpfRankingsParams = {
  period?: string;
  limit?: number;
};

export type GetMpfCategoriesParams = {
  period?: string;
};

export type GetMpfTrusteesParams = {
  period?: string;
};