# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: MPF Compare API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: mpf
    description: MPF fund data
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /mpf/funds:
    get:
      operationId: listMpfFunds
      tags: [mpf]
      summary: List all MPF funds with returns for a period
      parameters:
        - name: period
          in: query
          required: false
          schema:
            type: string
            default: "1y"
        - name: trustee
          in: query
          required: false
          schema:
            type: string
        - name: category
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of funds
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MpfFundSummary"

  /mpf/funds/rankings:
    get:
      operationId: getMpfRankings
      tags: [mpf]
      summary: Get top and bottom fund rankings for a period
      parameters:
        - name: period
          in: query
          required: false
          schema:
            type: string
            default: "1y"
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Rankings response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MpfRankingsResponse"

  /mpf/funds/{cfId}:
    get:
      operationId: getMpfFund
      tags: [mpf]
      summary: Get individual fund detail with all returns
      parameters:
        - name: cfId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Fund detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MpfFundDetail"
        "404":
          description: Fund not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /mpf/categories:
    get:
      operationId: getMpfCategories
      tags: [mpf]
      summary: Get category average returns for a period
      parameters:
        - name: period
          in: query
          required: false
          schema:
            type: string
            default: "1y"
      responses:
        "200":
          description: Category stats
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MpfCategoryStat"

  /mpf/trustees:
    get:
      operationId: getMpfTrustees
      tags: [mpf]
      summary: Get trustee average returns for a period
      parameters:
        - name: period
          in: query
          required: false
          schema:
            type: string
            default: "1y"
      responses:
        "200":
          description: Trustee stats
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MpfTrusteeStat"

  /mpf/sync:
    post:
      operationId: triggerMpfSync
      tags: [mpf]
      summary: Trigger a manual data sync from MPFA
      responses:
        "200":
          description: Sync result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MpfSyncResult"

  /mpf/sync/status:
    get:
      operationId: getMpfSyncStatus
      tags: [mpf]
      summary: Get last sync status
      responses:
        "200":
          description: Sync status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MpfSyncStatus"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    MpfFundReturn:
      type: object
      properties:
        period:
          type: string
        annualizedReturn:
          type: ["number", "null"]
        cumulativeReturn:
          type: ["number", "null"]
      required:
        - period
        - annualizedReturn
        - cumulativeReturn

    MpfFundSummary:
      type: object
      properties:
        cfId:
          type: string
        nameEn:
          type: string
        nameZh:
          type: ["string", "null"]
        trustee:
          type: string
        trusteeCode:
          type: string
        scheme:
          type: string
        fundType:
          type: string
        fundCategory:
          type: string
        launchDate:
          type: ["string", "null"]
        fundSizeHkm:
          type: ["number", "null"]
        riskClass:
          type: ["integer", "null"]
        ferPct:
          type: ["number", "null"]
        returnAnn:
          type: ["number", "null"]
        returnCum:
          type: ["number", "null"]
        period:
          type: string
      required:
        - cfId
        - nameEn
        - trustee
        - trusteeCode
        - scheme
        - fundType
        - fundCategory
        - period

    MpfFundDetail:
      type: object
      properties:
        cfId:
          type: string
        nameEn:
          type: string
        nameZh:
          type: ["string", "null"]
        trustee:
          type: string
        trusteeCode:
          type: string
        scheme:
          type: string
        fundType:
          type: string
        fundCategory:
          type: string
        launchDate:
          type: ["string", "null"]
        fundSizeHkm:
          type: ["number", "null"]
        riskClass:
          type: ["integer", "null"]
        ferPct:
          type: ["number", "null"]
        mgmtFee:
          type: ["string", "null"]
        returns:
          type: array
          items:
            $ref: "#/components/schemas/MpfFundReturn"
      required:
        - cfId
        - nameEn
        - trustee
        - trusteeCode
        - scheme
        - fundType
        - fundCategory
        - returns

    MpfRankingsResponse:
      type: object
      properties:
        period:
          type: string
        asOf:
          type: string
        top:
          type: array
          items:
            $ref: "#/components/schemas/MpfFundSummary"
        bottom:
          type: array
          items:
            $ref: "#/components/schemas/MpfFundSummary"
      required:
        - period
        - asOf
        - top
        - bottom

    MpfCategoryStat:
      type: object
      properties:
        category:
          type: string
        avgReturn:
          type: ["number", "null"]
        fundCount:
          type: integer
        period:
          type: string
      required:
        - category
        - avgReturn
        - fundCount
        - period

    MpfTrusteeStat:
      type: object
      properties:
        trustee:
          type: string
        trusteeCode:
          type: string
        avgReturn:
          type: ["number", "null"]
        fundCount:
          type: integer
        period:
          type: string
        funds:
          type: array
          items:
            $ref: "#/components/schemas/MpfFundSummary"
      required:
        - trustee
        - trusteeCode
        - avgReturn
        - fundCount
        - period
        - funds

    MpfSyncResult:
      type: object
      properties:
        message:
          type: string
        count:
          type: integer
        error:
          type: ["string", "null"]
      required:
        - message
        - count

    MpfSyncStatus:
      type: object
      properties:
        status:
          type: string
        fundsScraped:
          type: ["integer", "null"]
        errorMessage:
          type: ["string", "null"]
        startedAt:
          type: string
        completedAt:
          type: ["string", "null"]
      required:
        - status
        - startedAt