# lib/api-zod/src/generated/types/mpfCategoryStat.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MPF Compare API
 * OpenAPI spec version: 0.1.0
 */

export interface MpfCategoryStat {
  category: string;
  /** @nullable */
  avgReturn: number | null;
  fundCount: number;
  period: string;
}