# lib/api-zod/src/generated/types/mpfFundDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MPF Compare API
 * OpenAPI spec version: 0.1.0
 */
import type { MpfFundReturn } from "./mpfFundReturn";

export interface MpfFundDetail {
  cfId: string;
  nameEn: string;
  /** @nullable */
  nameZh?: string | null;
  trustee: string;
  trusteeCode: string;
  scheme: string;
  fundType: string;
  fundCategory: string;
  /** @nullable */
  launchDate?: string | null;
  /** @nullable */
  fundSizeHkm?: number | null;
  /** @nullable */
  riskClass?: number | null;
  /** @nullable */
  ferPct?: number | null;
  /** @nullable */
  mgmtFee?: string | null;
  returns: MpfFundReturn[];
}