# lib/api-zod/src/generated/types/mpfFundReturn.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MPF Compare API
 * OpenAPI spec version: 0.1.0
 */

export interface MpfFundReturn {
  period: string;
  /** @nullable */
  annualizedReturn: number | null;
  /** @nullable */
  cumulativeReturn: number | null;
}