# lib/api-zod/src/generated/types/mpfRankingsResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MPF Compare API
 * OpenAPI spec version: 0.1.0
 */
import type { MpfFundSummary } from "./mpfFundSummary";

export interface MpfRankingsResponse {
  period: string;
  asOf: string;
  top: MpfFundSummary[];
  bottom: MpfFundSummary[];
}