# lib/api-zod/src/generated/types/mpfSyncResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MPF Compare API
 * OpenAPI spec version: 0.1.0
 */

export interface MpfSyncResult {
  message: string;
  count: number;
  /** @nullable */
  error?: string | null;
}