# lib/api-zod/src/generated/types/mpfSyncStatus.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MPF Compare API
 * OpenAPI spec version: 0.1.0
 */

export interface MpfSyncStatus {
  status: string;
  /** @nullable */
  fundsScraped?: number | null;
  /** @nullable */
  errorMessage?: string | null;
  startedAt: string;
  /** @nullable */
  completedAt?: string | null;
}