# lib/api-zod/src/generated/types/mpfTrusteeStat.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MPF Compare API
 * OpenAPI spec version: 0.1.0
 */
import type { MpfFundSummary } from "./mpfFundSummary";

export interface MpfTrusteeStat {
  trustee: string;
  trusteeCode: string;
  /** @nullable */
  avgReturn: number | null;
  fundCount: number;
  period: string;
  funds: MpfFundSummary[];
}